+++ recipes/default.rb
@@ -1,27 +1,16 @@
+# Create tmp directories
+directory "#{node['rails-stack']['data_path']}/tmp" do
+  recursive true
+  mode '0777'
+end
+
 include_recipe 'rails-stack::update_packages'
 
 # User deployer
 include_recipe 'rails-stack::user'
 include_recipe 'rails-stack::lib_directory'
 
-# Rbenv
-include_recipe 'rbenv'
-include_recipe 'rbenv::ruby_build'
-include_recipe 'rbenv::rbenv_vars'
-
-directory "#{node['rails-stack']['data_path']}/tmp" do
-  recursive true
-  mode '0777'
-end
-
-ENV['TMPDIR'] = "#{node['rails-stack']['data_path']}/tmp"
-rbenv_ruby node[:ruby][:version]
-
-rbenv_gem 'bundler' do
-  ruby_version node[:ruby][:version]
-end
-
-include_recipe 'rails-stack::rbenv_wrapper'
+include_recipe 'rails-stack::ruby'
 
 # Monit
 include_recipe 'monit'