+++ recipes/rbenv_wrapper.rb
@@ -8,4 +8,3 @@ template "#{node['rails-stack']['data_path']}/bin/rbenv-exec" do
   source 'rbenv/wrapper.sh.erb'
   mode '0755'
 end
-