+++ recipes/update_packages.rb
@@ -3,7 +3,8 @@ general_packages = case node['platform_family']
                      include_recipe 'apt'
                      %w(libreadline6 libreadline6-dev git autoconf libssl-dev zlib1g zlib1g-dev libssl-dev libyaml-dev libsqlite3-dev sqlite3 libxml2-dev libxslt-dev libpq-dev)
                    when 'rhel'
-                     %w(git openssl-devel libyaml-devel readline readline-devel zlib-devel gcc curl-devel httpd-devel apr-devel apr-util-devel sqlite-devel libxml2-devel libxslt-devel )
+                     include_recipe 'yum-epel'
+                     %w(git openssl-devel readline readline-devel zlib-devel gcc curl-devel httpd-devel apr-devel apr-util-devel sqlite-devel libxml2-devel libxslt-devel )
                    else
                      raise node['platform_family']
                    end