+++ recipes/user.rb
@@ -37,6 +37,6 @@ end
 
 sudo deployer do
   user deployer
-  commands ['/usr/bin/monit', '/etc/init.d/nginx', '/etc/init.d/monit']
+  commands ['/usr/bin/monit', '/etc/init.d/nginx', '/etc/init.d/monit', '/usr/bin/tail -f /var/log/messages']
   nopasswd true
 end